rag-server/requirements.txt
flask==3.0.0
flask-cors==4.0.0
google-generativeai==0.3.2
langchain==0.1.0
langchain-community==0.0.13
langchain-huggingface==0.0.1
faiss-cpu==1.7.4
python-dotenv==1.0.0
sentence-transformers==2.2.2